Remove pairs from EVM chain tokelists.json (#17637)
* Remove pairs from EVM chain tokelists.json

* Add tokenlist-extended.json

* Add command for adding token to tokenlist-extended.json

* Remove empty pairs

* Update config namings

* Add command for adding token to tokenlist.json

* Update config and file namings

* Update tokenlist-extended.json

* Add token info getting

* Fix

* Fix

* Add duplicates check

* Add duplicate check in token list

* Update commands.go
