+++ .github/assets.config.yaml
@@ -1,14 +1,17 @@
 app:
-  log_level: debug
+  log_level: "debug"
 
 client_urls:
   binance:
-    dex: https://dex.binance.org
-    explorer: https://explorer.binance.org
-  assets_manager_api: "https://assets-manager-api.herokuapp.com"
+    dex: "https://dex.binance.org"
+    explorer: "https://explorer.binance.org"
+  assets_manager_api: "https://api.assets.trustwallet.com"
 
 urls:
-  tw_assets_app: https://assets.trustwalletapp.com
+  assets_app: "https://assets.trustwalletapp.com"
+  logo: "https://trustwallet.com/assets/images/favicon.png"
+
+time_format: "2006-01-02T15:04:05.000000"
 
 validators_settings:
   root_folder:
@@ -45,6 +48,7 @@ validators_settings:
     allowed_files:
       - "assets"
       - "tokenlist.json"
+      - "tokenlist-extended.json"
       - "validators"
       - "info"
