+++ blockchains/polygon/tokenlist-extended.json
@@ -1,0 +1,12 @@
+{
+    "name": "Trust Wallet: Polygon Extended List",
+    "logoURI": "https://trustwallet.com/assets/images/favicon.png",
+    "timestamp": "2022-01-27T14:33:26.183301",
+    "tokens": [
+    ],
+    "version": {
+        "major": 1,
+        "minor": 0,
+        "patch": 0
+    }
+}
