+++ blockchains/polygon/tokenlist.json
@@ -10,30 +10,7 @@
             "name": "Wrapped Ether",
             "symbol": "WETH",
             "decimals": 18,
-            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0x7ceB23fD6bC0adD59E62ac25578270cFf1b9f619/logo.png",
-            "pairs": [
-                {
-                    "base": "c966_t0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270"
-                },
-                {
-                    "base": "c966_t0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174"
-                },
-                {
-                    "base": "c966_t0x53E0bca35eC356BD5ddDFebbD1Fc0fD03FaBad39"
-                },
-                {
-                    "base": "c966_t0x831753DD7087CaC61aB5644b308642cc1c33Dc13"
-                },
-                {
-                    "base": "c966_t0x8f3Cf7ad23Cd3CaDbD9735AFf958023239c6A063"
-                },
-                {
-                    "base": "c966_t0xc2132D05D31c914a87C6611C10748AEb04B58e8F"
-                },
-                {
-                    "base": "c966_t0xD6DF932A45C0f255f85145f286eA0b292B21C90B"
-                }
-            ]
+            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0x7ceB23fD6bC0adD59E62ac25578270cFf1b9f619/logo.png"
         },
         {
             "asset": "c966_t0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270",
@@ -42,21 +19,7 @@
             "name": "Wrapped MATIC",
             "symbol": "WMATIC",
             "decimals": 18,
-            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270/logo.png",
-            "pairs": [
-                {
-                    "base": "c966_t0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174"
-                },
-                {
-                    "base": "c966_t0x7ceB23fD6bC0adD59E62ac25578270cFf1b9f619"
-                },
-                {
-                    "base": "c966_t0x831753DD7087CaC61aB5644b308642cc1c33Dc13"
-                },
-                {
-                    "base": "c966_t0xc2132D05D31c914a87C6611C10748AEb04B58e8F"
-                }
-            ]
+            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270/logo.png"
         },
         {
             "asset": "c966_t0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174",
@@ -65,18 +28,7 @@
             "name": "USD Coin (PoS)",
             "symbol": "USDC",
             "decimals": 6,
-            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174/logo.png",
-            "pairs": [
-                {
-                    "base": "c966_t0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270"
-                },
-                {
-                    "base": "c966_t0x7ceB23fD6bC0adD59E62ac25578270cFf1b9f619"
-                },
-                {
-                    "base": "c966_t0xc2132D05D31c914a87C6611C10748AEb04B58e8F"
-                }
-            ]
+            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174/logo.png"
         },
         {
             "asset": "c966_t0xc2132D05D31c914a87C6611C10748AEb04B58e8F",
@@ -85,18 +37,7 @@
             "name": "(PoS) Tether USD",
             "symbol": "USDT",
             "decimals": 6,
-            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0xc2132D05D31c914a87C6611C10748AEb04B58e8F/logo.png",
-            "pairs": [
-                {
-                    "base": "c966_t0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270"
-                },
-                {
-                    "base": "c966_t0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174"
-                },
-                {
-                    "base": "c966_t0x7ceB23fD6bC0adD59E62ac25578270cFf1b9f619"
-                }
-            ]
+            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0xc2132D05D31c914a87C6611C10748AEb04B58e8F/logo.png"
         },
         {
             "asset": "c966_t0x8f3Cf7ad23Cd3CaDbD9735AFf958023239c6A063",
@@ -105,12 +46,7 @@
             "name": "(PoS) Dai Stablecoin",
             "symbol": "DAI",
             "decimals": 18,
-            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0x8f3Cf7ad23Cd3CaDbD9735AFf958023239c6A063/logo.png",
-            "pairs": [
-                {
-                    "base": "c966_t0x7ceB23fD6bC0adD59E62ac25578270cFf1b9f619"
-                }
-            ]
+            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0x8f3Cf7ad23Cd3CaDbD9735AFf958023239c6A063/logo.png"
         },
         {
             "asset": "c966_t0x53E0bca35eC356BD5ddDFebbD1Fc0fD03FaBad39",
@@ -119,8 +55,7 @@
             "name": "ChainLink Token",
             "symbol": "LINK",
             "decimals": 18,
-            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0x53E0bca35eC356BD5ddDFebbD1Fc0fD03FaBad39/logo.png",
-            "pairs": []
+            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0x53E0bca35eC356BD5ddDFebbD1Fc0fD03FaBad39/logo.png"
         },
         {
             "asset": "c966_t0x831753DD7087CaC61aB5644b308642cc1c33Dc13",
@@ -129,8 +64,7 @@
             "name": "QuickSwap",
             "symbol": "QUICK",
             "decimals": 18,
-            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0x831753DD7087CaC61aB5644b308642cc1c33Dc13/logo.png",
-            "pairs": []
+            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0x831753DD7087CaC61aB5644b308642cc1c33Dc13/logo.png"
         },
         {
             "asset": "c966_t0xD6DF932A45C0f255f85145f286eA0b292B21C90B",
@@ -139,8 +73,7 @@
             "name": "Aave (PoS)",
             "symbol": "AAVE",
             "decimals": 18,
-            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0xD6DF932A45C0f255f85145f286eA0b292B21C90B/logo.png",
-            "pairs": []
+            "logoURI": "https://assets.trustwalletapp.com/blockchains/polygon/assets/0xD6DF932A45C0f255f85145f286eA0b292B21C90B/logo.png"
         }
     ],
     "version": {
