+++ blockchains/smartchain/tokenlist-extended.json
@@ -1,0 +1,13 @@
+{
+    "name": "Trust Wallet: SmartChain Extended List",
+    "logoURI": "https://trustwallet.com/assets/images/favicon.png",
+    "timestamp": "2022-01-27T14:33:26.183301",
+    "tokens": [
+        
+    ],
+    "version": {
+        "major": 1,
+        "minor": 0,
+        "patch": 0
+    }
+}
