+++ internal/config/config.go
@@ -11,6 +11,7 @@ type (
 		App                App                `mapstructure:"app"`
 		ClientURLs         ClientURLs         `mapstructure:"client_urls"`
 		URLs               URLs               `mapstructure:"urls"`
+		TimeFormat         string             `mapstructure:"time_format"`
 		ValidatorsSettings ValidatorsSettings `mapstructure:"validators_settings"`
 	}
 
@@ -27,7 +28,8 @@ type (
 	}
 
 	URLs struct {
-		TWAssetsApp string `mapstructure:"tw_assets_app"`
+		AssetsApp string `mapstructure:"assets_app"`
+		Logo      string `mapstructure:"logo"`
 	}
 
 	ValidatorsSettings struct {
