+++ internal/processor/updaters_auto.go
@@ -32,9 +32,6 @@ const (
 	marketPairsLimit = 1000
 	tokensListLimit  = 10000
 
-	twLogoURL       = "https://trustwallet.com/assets/images/favicon.png"
-	timestampFormat = "2006-01-02T15:04:05.000000"
-
 	activeStatus = "active"
 )
 
@@ -139,7 +136,7 @@ func createInfoJSON(chain coin.Coin, a explorer.Bep2Asset) error {
 }
 
 func createTokenListJSON(chain coin.Coin, tokens []tokenlist.Token) error {
-	tokenListPath := path.GetTokenListPath(chain.Handle)
+	tokenListPath := path.GetTokenListPath(chain.Handle, path.TokenlistDefault)
 
 	var oldTokenList tokenlist.Model
 	err := fileLib.ReadJSONFile(tokenListPath, &oldTokenList)
@@ -160,8 +157,8 @@ func createTokenListJSON(chain coin.Coin, tokens []tokenlist.Token) error {
 
 	return fileLib.CreateJSONFile(tokenListPath, &tokenlist.Model{
 		Name:      fmt.Sprintf("Trust Wallet: %s", coin.Coins[chain.ID].Name),
-		LogoURI:   twLogoURL,
-		Timestamp: time.Now().Format(timestampFormat),
+		LogoURI:   config.Default.URLs.Logo,
+		Timestamp: time.Now().Format(config.Default.TimeFormat),
 		Tokens:    tokens,
 		Version:   tokenlist.Version{Major: oldTokenList.Version.Major + 1},
 	})
@@ -315,8 +312,8 @@ func getTokenType(symbol string, nativeCoinSymbol string, tokenType types.TokenT
 
 func getLogoURI(id, githubChainFolder, nativeCoinSymbol string) string {
 	if id == nativeCoinSymbol {
-		return path.GetChainLogoURL(config.Default.URLs.TWAssetsApp, githubChainFolder)
+		return path.GetChainLogoURL(config.Default.URLs.AssetsApp, githubChainFolder)
 	}
 
-	return path.GetAssetLogoURL(config.Default.URLs.TWAssetsApp, githubChainFolder, id)
+	return path.GetAssetLogoURL(config.Default.URLs.AssetsApp, githubChainFolder, id)
 }
